backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
aiosqlite==0.20.0
PyMuPDF>=1.26.7
pdfplumber==0.11.0
python-docx==1.1.0
pytesseract==0.3.13
scikit-learn==1.5.0
scipy==1.14.0
numpy==1.26.4
websockets==12.0
pydantic==2.9.0
pydantic-settings==2.5.0
httpx>=0.25.0
cryptography>=44.0.0
certifi>=2024.12.14

# Dev / Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
